preventing conversion warnings

+++ c++ gdiplus/graphics.hpp
@@ -137,37 +137,55 @@ public:
 		static
 		Rect inflate( const Rect &r, const auto _d )
 		{
-			return { r.x - _d, r.y - _d, r.width + _d * 2, r.height + _d * 2 };
+			return { r.x - static_cast<FLOAT>(_d),
+			         r.y - static_cast<FLOAT>(_d),
+			         r.width + static_cast<FLOAT>(_d * 2),
+			         r.height + static_cast<FLOAT>(_d * 2) };
 		}
 
 		static
 		Rect inflate( const Rect &r, const auto _dx, const auto _dy )
 		{
-			return { r.x - _dx, r.y - _dy, r.width + _dx * 2, r.height + _dy * 2 };
+			return { r.x - static_cast<FLOAT>(_dx),
+			         r.y - static_cast<FLOAT>(_dy),
+			         r.width + static_cast<FLOAT>(_dx * 2),
+			         r.height + static_cast<FLOAT>(_dy * 2) };
 		}
 
 		static
 		Rect inflate( const Rect &r, const auto _dx, const auto _dy, const auto _dw, const auto _dh )
 		{
-			return { r.x - _dx, r.y - _dy, r.width + _dx + _dw, r.height + _dy + _dh };
+			return { r.x - static_cast<FLOAT>(_dx),
+			         r.y - static_cast<FLOAT>(_dy),
+			         r.width + static_cast<FLOAT>(_dx) + static_cast<FLOAT>(_dw),
+			         r.height + static_cast<FLOAT>(_dy) + static_cast<FLOAT>(_dh) };
 		}
 
 		static
 		Rect deflate( const Rect &r, const auto _d )
 		{
-			return { r.x + _d, r.y + _d, r.width - _d * 2, r.height - _d * 2 };
+			return { r.x + static_cast<FLOAT>(_d),
+			         r.y + static_cast<FLOAT>(_d),
+			         r.width - static_cast<FLOAT>(_d * 2),
+			         r.height - static_cast<FLOAT>(_d * 2) };
 		}
 
 		static
 		Rect deflate( const Rect &r, const auto _dx, const auto _dy )
 		{
-			return { r.x + _dx, r.y + _dy, r.width - _dx * 2, r.height - _dy * 2 };
+			return { r.x + static_cast<FLOAT>(_dx),
+			         r.y + static_cast<FLOAT>(_dy),
+			         r.width - static_cast<FLOAT>(_dx * 2),
+			         r.height - static_cast<FLOAT>(_dy * 2) };
 		}
 
 		static
 		Rect deflate( const Rect &r, const auto _dx, const auto _dy, const auto _dw, const auto _dh )
 		{
-			return { r.x + _dx, r.y + _dy, r.width - _dx - _dw, r.height - _dy - _dh };
+			return { r.x + static_cast<FLOAT>(_dx),
+			         r.y + static_cast<FLOAT>(_dy),
+			         r.width - static_cast<FLOAT>(_dx) - static_cast<FLOAT>(_dw),
+			         r.height - static_cast<FLOAT>(_dy) - static_cast<FLOAT>(_dh) };
 		}
 
 		bool contains( const auto _x, const auto _y ) const
